Add files via upload
Economics 153 was a class I thoroughly enjoyed. It was a class that taught the foundations in the Labor Market. I gained a lot of insight and knowledge and used R as my primary software to help me better visualize the data presented. This folder showcases my skills and knowledge to turn raw data into graphs and visuals.